+++ packages/isar/tool/get_version.dart
@@ -1,0 +1,5 @@
+import 'package:isar/src/version.dart';
+
+void main() {
+  print(isarCoreVersion);
+}