+++ packages/isar_generator/lib/src/collection_generator.dart
@@ -25,7 +25,7 @@ class IsarCollectionGenerator extends GeneratorForAnnotation<Collection> {
     final object = IsarAnalyzer().analyze(element);
 
     var code = '''
-    // ignore_for_file: duplicate_ignore, non_constant_identifier_names, constant_identifier_names, invalid_use_of_protected_member, unnecessary_cast, unused_local_variable
+    // ignore_for_file: duplicate_ignore, non_constant_identifier_names, constant_identifier_names, invalid_use_of_protected_member, unnecessary_cast, unused_local_variable, no_leading_underscores_for_local_identifiers
 
     extension Get${object.dartName}Collection on Isar {
       IsarCollection<${object.dartName}> get ${object.accessor} => getCollection();